Disable shortcuts that I no longer use
Some of them are confusing (such as ^W, which is used to type window
commands).
Some are not used anymore because F1-F12 keys are either not present
(Chromebook) or usefully bound to other actions (backlight, sound
control etc).
Others are not compatible with some platforms and it's so damn hard to
properly disable system shortcuts, such as Ctrl+Up in Macbook.

Using vim commands looks prettier with every platform switch :)